Revert "change amount type of invoice creation to float (#217)"
This reverts commit 5827b2c1a055d86c6eaef7bfa9419c707ef94584.

+++ raiden/utils/typing.py
@@ -105,7 +105,7 @@ PaymentHashInvoice = NewType("PaymentHashInvoice", T_PaymentHashInvoice)
 T_PaymentAmount = int
 PaymentAmount = NewType("PaymentAmount", T_PaymentAmount)
 
-T_InvoiceAmount = float
+T_InvoiceAmount = int
 InvoiceAmount = NewType("InvoiceAmount", T_InvoiceAmount)
 
 T_Invoice_Expires = int
